Move Configuration cog code from lib/bot/__init__.py to lib/cogs/configuration.py
